hopefully no more hardcoded classes in detectron2.predict

+++ drone_detector/engines/detectron2/predict.py
@@ -31,8 +31,9 @@ def predict_bboxes_detectron2(path_to_model_config:Param("Path to pretrained mod
                               processing_dir:Param("Directory to save the intermediate tiles. Deleted after use", type=str, default='temp'),
                               tile_size:Param("Tile size to use. Default 400x400px tiles", type=int, default=400),
                               tile_overlap:Param("Tile overlap to use. Default 100px", type=int, default=200),
-                              use_tta:Param("Use test-time augmentation", store_true),
-                              smooth_preds:Param("Run fill_holes and dilate_erode to masks", store_true)
+                              use_tta:Param("Use test-time augmentation", store_false),
+                              postproc_results:Param('Filter predicted masks', store_true),
+                              smooth_preds:Param("Run fill_holes and dilate_erode to masks", store_false)
     ):
     "Detect bounding boxes from a new image using a pretrained model"
 
@@ -83,12 +84,13 @@ def predict_bboxes_detectron2(path_to_model_config:Param("Path to pretrained mod
         #preds = dilate_erode(preds)
 
     preds_coco = detectron2_bbox_preds_to_coco_anns(images, preds)
-    # TODO fix categories to not be hardcoded
-    preds_coco['categories'] = [
-        {'supercategory':'deadwood', 'id':1, 'name': 'uprightwood'},
-        {'supercategory':'deadwood', 'id':2, 'name': 'groundwood'},
-    ]
-
+    try:
+        with open(coco_set) as coco:
+            coco = json.load(coco)['categories']
+    except:
+        print('No categories found, defaulting to dummy classes')
+        cats = [{'supercategory': 'dummy', 'id':i+1, 'name': f'class_{i+1}'}
+                for i in range(cfg.ROI_HEADS.NUM_CLASSES)]
 
 
     # Process preds to shapefiles
@@ -100,29 +102,28 @@ def predict_bboxes_detectron2(path_to_model_config:Param("Path to pretrained mod
 
     coco_proc.coco_to_shp(preds_coco, downsample_factor=1)
 
-    # Discard all predictions that are not completely within -2m erosion of grid cell
-    # TODO add as optional postprocessing step
 
     grid = tiler.grid
     grid = grid.to_crs('EPSG:3067')
 
-    # Drop all polygons whose centroid is not within thresholded cell.
-    # Thresholded cell is constructed by eroding it by half the overlap area.
-    for cell in grid.itertuples():
-        if not os.path.isfile(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson'): continue
-        pred_shp = gpd.read_file(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
-        orig_crs = pred_shp.crs
-        pred_shp = pred_shp.to_crs('EPSG:3067')
-        cell_px = (cell.geometry.bounds[2] - cell.geometry.bounds[0]) / tile_size
-        cell_geom = cell.geometry.buffer(-(cell_px * overlap/2))
-        pred_shp['to_drop'] = pred_shp.apply(lambda row: 0 if row.geometry.centroid.within(cell_geom) else 1, axis=1)
-        pred_shp = pred_shp[pred_shp.to_drop == 0]
-        pred_shp = pred_shp.to_crs(orig_crs)
-        pred_shp.drop(columns=['to_drop'], inplace=True)
-        if use_tta:
-            pred_shp = do_nms(pred_shp, 0.1, 'score')
-        if len(pred_shp) > 0: pred_shp.to_file(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
-        else: os.remove(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
+    if postproc_results:
+        # Drop all polygons whose centroid is not within thresholded cell.
+        # Thresholded cell is constructed by eroding it by half the overlap area.
+        for cell in grid.itertuples():
+            if not os.path.isfile(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson'): continue
+            pred_shp = gpd.read_file(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
+            orig_crs = pred_shp.crs
+            pred_shp = pred_shp.to_crs('EPSG:3067')
+            cell_px = (cell.geometry.bounds[2] - cell.geometry.bounds[0]) / tile_size
+            cell_geom = cell.geometry.buffer(-(cell_px * overlap/2))
+            pred_shp['to_drop'] = pred_shp.apply(lambda row: 0 if row.geometry.centroid.within(cell_geom) else 1, axis=1)
+            pred_shp = pred_shp[pred_shp.to_drop == 0]
+            pred_shp = pred_shp.to_crs(orig_crs)
+            pred_shp.drop(columns=['to_drop'], inplace=True)
+            if use_tta:
+                pred_shp = do_nms(pred_shp, 0.1, 'score')
+            if len(pred_shp) > 0: pred_shp.to_file(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
+            else: os.remove(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
 
     # Collate shapefiles
     untile_vector(path_to_targets=f'{processing_dir}/predicted_vectors', outpath=outfile)
@@ -140,8 +141,11 @@ def predict_instance_masks_detectron2(path_to_model_config:Param("Path to pretra
                                       processing_dir:Param("Directory to save the intermediate tiles. Deleted after use", type=str, default='temp'),
                                       tile_size:Param("Tile size to use. Default 400x400px tiles", type=int, default=400),
                                       tile_overlap:Param("Tile overlap to use. Default 100px", type=int, default=200),
+                                      coco_set:Param('Path to json file for the COCO dataset the format was trained on',
+                                                     type=str),
                                       use_tta:Param("Use test-time augmentation", store_false),
-                                      smooth_preds:Param("Run fill_holes and dilate_erode to masks", store_true)
+                                      postproc_results:Param('Filter predicted masks', store_true),
+                                      smooth_preds:Param("Run fill_holes and dilate_erode to masks", store_false)
     ):
     "Segment instance masks from a new image using a pretrained model"
 
@@ -191,11 +195,13 @@ def predict_instance_masks_detectron2(path_to_model_config:Param("Path to pretra
         #preds = dilate_erode(preds)
 
     preds_coco = detectron2_mask_preds_to_coco_anns(images, preds)
-    # TODO fix categories to not be hardcoded
-    preds_coco['categories'] = [
-        {'supercategory':'deadwood', 'id':1, 'name': 'uprightwood'},
-        {'supercategory':'deadwood', 'id':2, 'name': 'groundwood'},
-    ]
+    try:
+        with open(coco_set) as coco:
+            coco = json.load(coco)['categories']
+    except:
+        print('No categories found, defaulting to dummy classes')
+        cats = [{'supercategory': 'dummy', 'id':i+1, 'name': f'class_{i+1}'}
+                for i in range(cfg.ROI_HEADS.NUM_CLASSES)]
 
     # Process preds to shapefiles
     coco_proc = COCOProcessor(data_path=processing_dir,
@@ -206,29 +212,29 @@ def predict_instance_masks_detectron2(path_to_model_config:Param("Path to pretra
 
     coco_proc.coco_to_shp(preds_coco, downsample_factor=1)
 
+    if postproc_results:
+        # Drop all polygons whose centroid is not within thresholded cell.
+        # Thresholded cell is constructed by eroding it by half the overlap area.
+        grid = tiler.grid
+        grid = grid.to_crs('EPSG:3067')
+        num_polys = 0
+        for cell in grid.itertuples():
+            if not os.path.isfile(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson'): continue
+            pred_shp = gpd.read_file(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
+            num_polys += len(pred_shp)
+            orig_crs = pred_shp.crs
+            pred_shp = pred_shp.to_crs('EPSG:3067')
+            cell_px = (cell.geometry.bounds[2] - cell.geometry.bounds[0]) / tile_size
+            cell_geom = cell.geometry.buffer(-(cell_px * tile_overlap/2))
+            pred_shp['to_drop'] = pred_shp.apply(lambda row: 0 if row.geometry.centroid.within(cell_geom) else 1, axis=1)
+            pred_shp = pred_shp[pred_shp.to_drop == 0]
+            pred_shp = pred_shp.to_crs(orig_crs)
+            pred_shp.drop(columns=['to_drop'], inplace=True)
+            if len(pred_shp) > 0: pred_shp.to_file(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
+            else: os.remove(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
 
-    # Drop all polygons whose centroid is not within thresholded cell.
-    # Thresholded cell is constructed by eroding it by half the overlap area.
-    grid = tiler.grid
-    grid = grid.to_crs('EPSG:3067')
-    num_polys = 0
-    for cell in grid.itertuples():
-        if not os.path.isfile(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson'): continue
-        pred_shp = gpd.read_file(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
-        num_polys += len(pred_shp)
-        orig_crs = pred_shp.crs
-        pred_shp = pred_shp.to_crs('EPSG:3067')
-        cell_px = (cell.geometry.bounds[2] - cell.geometry.bounds[0]) / tile_size
-        cell_geom = cell.geometry.buffer(-(cell_px * tile_overlap/2))
-        pred_shp['to_drop'] = pred_shp.apply(lambda row: 0 if row.geometry.centroid.within(cell_geom) else 1, axis=1)
-        pred_shp = pred_shp[pred_shp.to_drop == 0]
-        pred_shp = pred_shp.to_crs(orig_crs)
-        pred_shp.drop(columns=['to_drop'], inplace=True)
-        if len(pred_shp) > 0: pred_shp.to_file(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
-        else: os.remove(f'{processing_dir}/predicted_vectors/{cell.cell}.geojson')
-
-    # Collate shapefiles
-    print(f'{num_polys} polygons before edge area removal')
+        # Collate shapefiles
+        print(f'{num_polys} polygons before edge area removal')
     untile_vector(path_to_targets=f'{processing_dir}/predicted_vectors', outpath=outfile)
 
     print('Removing intermediate files')
